+++ src/main/java/com/piseth/java/school/phoneshopenight/exception/ResourceNotFoundException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 
 public class ResourceNotFoundException extends ApiException{
 	
-	public ResourceNotFoundException(String resourceName, Integer id) {
+	public ResourceNotFoundException(String resourceName, Long id) {
 		super(HttpStatus.NOT_FOUND, String.format("%s With id = %d not found",resourceName,id ));
 	}
